pyproject.toml
[tool.poetry]
name = "test_medicine"
version = "0.1.0"
description = ""
authors = ["Artem Protsenko"]

[tool.poetry.dependencies]
python = "^3.7"
django = "3.2.*"
python-dotenv = "0.18.*"
psycopg2-binary = "2.8.*"
djangorestframework = "3.12.*"
drf-yasg = "1.20.*"
django-rest-swagger = "2.2.*"

[tool.poetry.group.dev.dependencies]
pytest-django = "3.8.*"
pytest = "5.3.*"
factory-boy = "^3.2.1"
responses = "0.10.*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
